set up for parsing with new parsing algorithm
Create a class that takes a testimony string and breaks it into a list
of contributions in the order they were spoken.

This class has a skeleton method to match the speaker's name to a name
in a metadata table. It is currently not implemented but a class that
loads the metadata information, supplemented with hearing attendance
data and string matching algorithms used in the
Parser._match_speaker_to_member_id() method should work well.

Currently, all parsed speakers will have a member_id set to 0.

The table schema has been updated with a table to store the new parsed
statements