New parser functions, switching development goal
At this point, I've discovered a bug in gpo_tools.scrape where the
summary information it collects is only from the first 100 hearings per
congress as returned by the API.

As a result, I'm freezing development on the parsing and speaker
matching components and switching to focus on a robust api client that
utilizes requests and sqlalchemy to better query and manage the hearing
summary data.

Once this is done, I will return to regression testing of the parsing
and speaker matching components.