refactor to extract package extraction method
The previous implementation of get_package_ids_by_congress contained the
general case for getting packages without the congressional parameter.
This logic has been extracted to its own method for greater utility.

Now that the param variable is built with the congress param first, the
test regex needed to be changed accordingly.