implement xml parsers, fix namespace bug
private methods that parse the mod page xml using xpath expressions are
implemented here. They return lists of dictionaries that will have to be
parsed further to create ORM objects to be added.

A database handler class that does this parsing will need to query
existing committee, member, and subcommittee records so that duplicate
insertions are not performed.

the namespace var in _get_mod_fields was reworked to be a dict to fix
the implementations of the xml parsers' uses of etree_root.xpath()