Change logger to be injected to fix bug.
The way the old logging setup worked was a logger for each instance of
the mods parsing class. This meant that every instance involved creating
a new logger that opened a file pointer to the same file. This file was
being opened many many times and overwriting all previous messages. It
also made it insanely slow. Now logging is done with a single logger
created in the APIClient constructor and passed into the mods xml parser
via the constructor.