add method for insert/update transcript.
sync_transcripts is a potentially over-complex way to add transcript
records to a new table in the database. Upon initialization of
db_handler instances, a cache of existing transcripts is created in the
form of a dict with package_id as a key and a 32-bit hash of the body
text of the transcript as the value.

The sync_transcripts method iterates over a dict of transcripts (keys
are again package_ids) and searches the cache for existing entries for
this hearing. Upon finding one, it checks if the hashes are equal. If
not, pull the transcript from the db and update it and update the cache.
If no existing records are found a new record is inserted into the db
and the cache.